* restructure files * add week 2 * add catch operators
