+++ app/src/main/java/com/infinite/rx/challenge/Utils.kt
@@ -1,4 +1,4 @@
-package com.infinite.rx.challenge.week1
+package com.infinite.rx.challenge
 
 
 fun String.print() {
